Separate the addItem function
* Create addItem.js to deal with the addElement to the ul list
* Change the html order and styles
