+++ src/index.html
@@ -10,8 +10,8 @@
     />
   </head>
   <body>
-    <ul id="todo-list">
+    <div id="box">
       <button class='link-button'>Clear all completed</button>
-    </ul>
+    </div>
   </body>
 </html>
